requirements.txt
langgraph
langchain
langchain-community
langchain-google-genai
python-dotenv
pandas
numpy
matplotlib
seaborn
scikit-learn
